requirements.txt
amqp==5.3.1
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.4
cassidy==0.1.4
celery==5.6.0
certifi==2026.2.25
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.6
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.1
daphne==4.2.1
dj-database-url==3.0.1
Django==5.2.5
django-cors-headers==4.7.0
django-silk==5.4.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
enumb==0.1.5
exceptiongroup==1.3.1
expo==0.1.2
gprof2dot==2025.4.14
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.3.0
kombu==5.6.1
mimetype==0.1.5
msgpack==1.1.1
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg==3.2.9
psycopg-binary==3.2.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.3.0
pytest==9.0.2
pytest-django==4.12.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
redis==6.4.0
requests==2.32.5
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.2.14
zope.interface==8.0
